+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to the "csv-table-viewer" extension will be documented in th
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
+## [0.0.2] - 2026-03-23
+
+- single column sort
+- hide/unhide columns
+
 ## [0.0.1] - 2026-03-23
 
 - Initial release